+++ README.md
@@ -1,2 +1,1 @@
-# Car-Rental issue #4 testcommit rajnaig
-koszi
+# Car-Rental
