--- proba.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Próba</title>
-</head>
-
-<body>
-  <h2>Tesztelés</h2>
-</body>
-
-</html>
